Finished algorithm for two contacts
Script reads partition.inp file and creates principle layers. It
is testing the created layers and also creates a filename.part.js
script for jmol where the atoms have a label and are colored. Next
steps would be to modify the algorithm for more contacts and
putting the testing into a standalone script.